mvp/photome-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic[email]==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
boto3==1.35.93
botocore==1.35.93
httpx==0.28.1
Pillow==11.1.0
python-dateutil==2.9.0
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
